requirements.txt
annotated-types==0.6.0
Authlib==1.3.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dimod==0.12.14
diskcache==5.6.3
dwave-cloud-client==0.11.3
dwave-greedy==0.3.0
dwave-hybrid==0.6.11
dwave-inspector==0.4.4
dwave-neal==0.6.0
dwave-networkx==0.8.14
dwave-ocean-sdk==6.9.0
dwave-preprocessing==0.6.5
dwave-samplers==1.2.0
dwave-system==1.23.0
dwave-tabu==0.5.0
dwavebinarycsp==0.3.0
fasteners==0.19
Flask==3.0.2
homebase==1.0.1
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
minorminer==0.2.13
networkx==3.2.1
numpy==1.26.4
packaging==24.0
penaltymodel==1.1.0
plucky==0.4.3
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
rectangle-packer==2.0.2
requests==2.31.0
scipy==1.13.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
